Removed reference to "no routing over overcrowding" if the new pathing system is enabled.
Changed some settings in the .vcproj file. If compiling manually, the "NEW PATHING" compiler option must be used to get the new pathing algorithm.

Credit limit is restored slowly after a period of indebtedness.

Revenue system updated to calculate the effect of the convoy's or line's speed on the likelihood of passengers using private cars.

The "avoid overcrowding" setting now causes passengers/goods to be discarded at any overcrowded intermediate stop, not being their destination, and not pay for the journey there.

A necessary file for the quickstone hashtable, missing from previous uploads, and necessary for the new pathing system, is included.