ADD: It is now possible to create low (half height) bridges over roads, tramways and waterways if: - one is the public player; - the lower way has a speed limit of zero (e.g., a stream; this was already possible); - the lower way is owned by the same player as the player building the bridge; or - the lower way is unowned and not a public right of way.
These low bridges cannot be passed by vehicles with a new is_tall=1 flag set in their .dat files (default: is_tall=0). Tall vehicles are stated to be such in the depot/replace windows. This is intended to be used with double-decker 'buses and trams, sailing ships, etc. The idea is to give single decker 'buses and trams a function in the game.

NOTE: This will need a recompile of Makeobj to allow tall vehicles to be created/recognised as such.