fix potential overflow in inauguration_time
inauguration_time is an unsigned 32 bit value but it's being set from
get_zeit_ms() which returns a 64bit signed value.  If the 64bit signed
value won't fit into the 32 bit type (>4294967291) then it will overflow
and wrap to 0.  This could lead to odd behaviour with industries not
registering with stops midway through a long running game.

Diagnosed here:
http://forum.simutrans.com/index.php?topic=16673.msg160669#msg160669