ADD: Different buildings can be specified to generate different proportions of passengers of different classes. By default, buildings will generate equal proportions of all classes. To specify proportions of classes, use the following format: class_proportion[0]=60 class_proportion[1]=30 class_proportion[2]=5 class_proportion[3]=4 class_proportion[4]=1
The numbers in square brackets are the class numbers, and the numbers after the = sign are percentages. In the above example, 60% of the passengers generated are class zero, 30% class 1 and so forth.

TODO: Apply the same principle to buildings that demand passengers (the above .dat file entries already work for such buildings to store the data, but this feature is not yet implemented in the code).
NOTE: This requires a makeobj upgrade to work.
CODE: Allow the debug build of makeobj to work