Fixing the layout in the upper part of the convoy assembler
Pass and mail:
Shows how many classes there currently are in the convoy along with how
much people/mail those can fit. The classes update when changed.
good:
The first four good types that you select by bying a vehicle are now
visible in the depot as a sum of all the good of a given type. The
layout is smart, so to speak that it senses how much space there is
(from the reassign class button and the class entries which have
priority) and sums up the best way possible what and how much good the
vehicle can hold.
Currently some of the labels have a small tendency to be too long for
the window, slightly sticking out over the edge. It is not a big deal,
but can be addressed if needed