A bunch of additions:
Add: Button in convoi_detail to open class manager
Add: in convoi_detail, freight is now kind of displayed pr class.. There
are multiple bugs, one being that it flickers quite much around. That
will be fixed later on when I know how to fix the following bug:
Add: Attempt to sort freight into compartments in simvehicle.cc. It
fails, and if no solution is found, I must abandon the idea
Add: Initial display of classes in freight_list_sorter (the 'thing' that
displays the freight info in convoy- and halt windows). This will be
build further upon, and are just there for testing currently. It should
have some more class entries and eventually also be possible to sort by
class
Add: in simvehicle.cc, a new routine "get_cargo_class_info(cbuffer_t &
buf, uint8 g_class)" to show what cargo is in the vehicle. It is
basically a duplication of "get_cargo_class_info(cbuffer_t & buf, uint8
g_class)", but it only sends passengers or mail from certain classes to
the freight list sorter, usefull for showing capacity per class in the
convoy detail window. It do however suffer from the same issues as the
third paragraph in this commit.