Fix: A new way of displaying the freight of a vehicle in convoy_detail
It now goes through the freight_list_sorter and gets sorted. Currently
the sorting list is hardcoded for passenger and mail to:
by_class_detail, and all other freight to: by_destination_detail.
This is, however, easily changeable, and one could even put a button in
the window to change sort option if desired.