Change SDL2 fullscreen mode to run in fullscreen desktop mode rather …
…than exclusive fullscreen mode.

There is an issue or bug with SDL2 exclusive fullscreen display mode changes, such as during minimize and maximize, that results in event polling calls blocking for around 3 seconds. This problem occurs each time display mode is changed so a minimize and maximize cycle would block for around 6 seconds. Event polling calls are not meant to block by design.

git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8674 8aca7d54-2c30-db11-9de9-000461428c89