Add an option to the Makefile for creating compile_command.sjson
`compile_command.json` is a compilation database files used by different
clangd based IDEs and editors for more accurate code completion.

In this patch a make target `compile-commands` is added as a PHONY
target which creates such a file. The target is PHONY for easy updating
of the database file.

The logic used for creating is trivial and will treat all files as C++
(even the `.m`, `.mm` and `.c` files) and won't do escaping of `"` in
any path.

Information on `compile_commands.json` can be found on
https://clang.llvm.org/docs/JSONCompilationDatabase.html